Refactor GraphQLService.cpp visitors into RequestLoader.*
